+++ client/src/test/java/com/influxdb/client/ITNotificationRulesApi.java
@@ -52,6 +52,7 @@ import com.influxdb.exceptions.NotFoundException;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
@@ -450,6 +451,7 @@ class ITNotificationRulesApi extends AbstractITClientTest {
     }
 
     @Test
+    @Disabled("TODO https://github.com/influxdata/influxdb/issues/18409")
     void labelDeleteNotExists() {
 
         SlackNotificationEndpoint endpoint = notificationEndpointsApi