+++ client/src/test/java/com/influxdb/client/ITTelegrafsApi.java
@@ -40,6 +40,7 @@ import com.influxdb.exceptions.NotFoundException;
 import com.moandjiezana.toml.Toml;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
@@ -369,6 +370,7 @@ class ITTelegrafsApi extends AbstractITClientTest {
     }
 
     @Test
+    @Disabled("TODO https://github.com/influxdata/influxdb/issues/18409")
     void labelDeleteNotExists() {
 
         Telegraf telegrafConfig = telegrafsApi