+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## 1.9.0 [unreleased]
 
-### API
+### Features
+1. [#119](https://github.com/influxdata/influxdb-client-java/pull/119): Scala and Kotlin clients has their own user agent string
 
+### API
 1. [#117](https://github.com/influxdata/influxdb-client-java/pull/117): Update swagger to latest version
 
 ### Bug Fixes
-
 1. [#116](https://github.com/influxdata/influxdb-client-java/pull/116): The closing message of the `WriteApi` has `Fine` log level
 
 ### Dependencies